refactor: move components into ui lib
