--- projects/app/src/app/common/components/form-field/components/form-field-label/form-field-label.component.html
@@ -1,11 +1,0 @@
-<label [attr.for]="context.id()">
-  <ng-content></ng-content>
-
-  @if (isRequired()) {
-    (*)
-  }
-
-  @if (isOptional()) {
-    (?)
-  }
-</label>
