--- projects/app/src/app/common/components/form-field/context.service.ts
@@ -1,6 +1,0 @@
-import { Injectable, signal } from '@angular/core';
-
-@Injectable()
-export class FormFieldContextService {
-  id = signal<string | null>(null);
-}
