--- projects/app/src/app/common/components/form-field/index.ts
@@ -1,16 +1,0 @@
-export * from './components/form-field/form-field.component';
-export * from './components/form-field-error/form-field-error.component';
-export * from './components/form-field-hint/form-field-hint.component';
-export * from './components/form-field-label/form-field-label.component';
-
-import { FormFieldComponent } from './components/form-field/form-field.component';
-import { FormFieldErrorComponent } from './components/form-field-error/form-field-error.component';
-import { FormFieldHintComponent } from './components/form-field-hint/form-field-hint.component';
-import { FormFieldLabelComponent } from './components/form-field-label/form-field-label.component';
-
-export const FORM_FIELD_EXPORTS = [
-  FormFieldComponent,
-  FormFieldErrorComponent,
-  FormFieldHintComponent,
-  FormFieldLabelComponent,
-];
