+++ projects/app/src/app/common/components/index.ts
@@ -1,5 +1,2 @@
-export * from './autocomplete';
 export * from './change-category-modal';
-export * from './form-field';
-export * from './textarea';
 export * from './user-credentials-form';
