+++ projects/app/src/app/features/list/components/item-form-modal/item-form-modal.component.ts
@@ -28,14 +28,14 @@ import {
   ModalHeaderDirective,
 } from '@ui/components/modal';
 import { TextInputComponent } from '@ui/components/text-input';
-
+import { TextareaComponent } from '@ui/components/textarea';
+import { FORM_FIELD_EXPORTS } from '@ui/components/form-field';
 import {
   AUTOCOMPLETE_EXPORTS,
   AutocompleteAsyncOptionsFn,
   AutocompleteOption,
-  FORM_FIELD_EXPORTS,
-  TextareaComponent,
-} from '@app/common/components';
+} from '@ui/components/autocomplete';
+
 import { DEFAULT_CATEGORY } from '@app/core/constants';
 import { UiStore } from '@app/core/ui';
 import { InventoryStore } from '@app/features/inventory/store';
