+++ projects/ui/src/components/autocomplete/autocomplete-option/autocomplete-option.component.ts
