+++ projects/ui/src/components/autocomplete/autocomplete/autocomplete.component.css
