+++ projects/ui/src/components/form-field/form-field-label/form-field-label.component.html
@@ -1,0 +1,5 @@
+<label [attr.for]="formField.id()">
+  <ng-content></ng-content>
+
+  @if (isRequired()) { (*) } @if (isOptional()) { (?) }
+</label>
