+++ projects/ui/src/components/public-api.ts
@@ -1,9 +1,11 @@
 export * from './actions-menu/public-api';
+export * from './autocomplete/public-api';
 export * from './bottom-menu/public-api';
 export * from './button/public-api';
 export * from './card-list/public-api';
 export * from './checkbox/public-api';
 export * from './confirm-prompt-modal/public-api';
+export * from './form-field/public-api';
 export * from './icon-button/public-api';
 export * from './linear-spinner/public-api';
 export * from './modal/public-api';
@@ -12,4 +14,5 @@ export * from './page-header/public-api';
 export * from './quick-number/public-api';
 export * from './select/public-api';
 export * from './text-input/public-api';
+export * from './textarea/public-api';
 export * from './toggle/public-api';
