+++ projects/ui/src/components/textarea/public-api.ts
